Support AARCH64 for Fuchsia
This change adds support for compiling Fuchsia for ARM64. It also
changes the arguments passed to build-script to remove hardcoded file
paths to sysroots and include directories in a Fuchsia build tree.

Some changes were made to be able to support building more than one
architecture for a non-Darwin platform as the Unix sdk targets only
support one at a time.